requirements.txt
blueprint==3.4.2
Flask==1.1.2
passlib==1.7.4
PyJWT==2.0.0
PyMySQL==0.10.1
requests==2.25.1
SQLAlchemy==1.3.22
SQLAlchemy-serializer==1.3.4.4
urllib3==1.26.2
Werkzeug==1.0.1
